move helpers to utils folder
