--- workspaces/orchestrator/e2e-tests/tests/specs/test-helpers.ts
@@ -1,1 +1,0 @@
-export * from "../support/utils/test-helpers.js";
